requirements.txt
astroid==2.8.0
attrs==21.2.0
black==21.9b0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
google-api-core==2.0.1
google-auth==2.2.1
google-cloud-bigquery==2.28.0
google-cloud-core==2.0.0
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
proto-plus==1.19.2
protobuf==3.18.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.11.1
pyparsing==2.4.7
pytest==6.2.5
regex==2021.9.30
requests==2.26.0
rsa==4.7.2
six==1.16.0
toml==0.10.2
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.6.0
